+++ lua/snailrace/game/logic.lua
@@ -1,10 +1,10 @@
 local driftwood = require("driftwood")
 
 local state = require("snailrace.state")
-local cards = require("snailrace.core.cards")
 local constants = require("snailrace.core.constants")
-local handlers = require("snailrace.core.effect_handlers")
 local utils = require("snailrace.core.utils")
+local cards = require("snailrace.cards.cards")
+local handlers = require("snailrace.cards.effect_handlers")
 local render = require("snailrace.game.render")
 local racers = require("snailrace.game.racers")
 
@@ -12,7 +12,7 @@ local logic = {}
 
 --- Check if all snails have finished the race.
 ---@param positions table<string, number> Snail positions.
----@return boolean True if all snails have finished, otherwise false.
+---@return boolean result True if all snails have finished, otherwise false.
 local function all_snails_finished(positions)
     for _, position in pairs(positions) do
         if position < constants.MAX_UNITS then
@@ -22,6 +22,42 @@ local function all_snails_finished(positions)
     return true
 end
 
+--- Rebuild and reshuffle the deck if required.
+---@param current_state table The current race state.
+---@return boolean result True if the race should finish due to exceeding redraw limits.
+local function handle_deck_rebuild(current_state)
+    if #current_state.deck <= current_state.deck_size_threshold then
+        driftwood.log.info("Deck size below threshold. Rebuilding and reshuffling.")
+        racers.build_deck(current_state)
+        current_state.redraws = current_state.redraws + 1
+        if current_state.redraws >= constants.MAX_REDRAW_ATTEMPTS then
+            driftwood.log.info("Max redraw limit reached. Stopping the race.")
+            return true
+        end
+    end
+    return false
+end
+
+--- Draw a valid card from the deck.
+---@param current_state table The current race state.
+---@return table|nil card The drawn card or nil if no valid cards remain.
+local function draw_valid_card(current_state)
+    local drawn_card
+    repeat
+        if #current_state.deck == 0 then
+            driftwood.log.info("Deck is empty. Rebuilding and reshuffling.")
+            racers.build_deck(current_state)
+            current_state.redraws = current_state.redraws + 1
+            if current_state.redraws >= constants.MAX_REDRAW_ATTEMPTS then
+                return nil
+            end
+        end
+        drawn_card = table.remove(current_state.deck, 1)
+    until not drawn_card or current_state.positions[drawn_card.snail_id] < constants.MAX_UNITS
+    return drawn_card
+end
+
+
 --- Perform a single race tick.
 logic.race_tick = function()
     local current_state = state.get()
@@ -30,43 +66,14 @@ logic.race_tick = function()
         return
     end
 
-    local finishing = false
-
-    -- Check if deck size is below threshold and rebuild if needed
-    if #current_state.deck <= current_state.deck_size_threshold then
-        driftwood.log.info("Deck size is below threshold. Rebuilding and reshuffling the deck.")
-        racers.build_deck(current_state)
-
-        current_state.redraws = current_state.redraws + 1
-        if current_state.redraws >= constants.MAX_REDRAW_ATTEMPTS then
-            driftwood.log.info("Max redraw limit reached. Stopping the race.")
-            finishing = true
-        end
+    -- Handle deck rebuild and check for finish conditions
+    if handle_deck_rebuild(current_state) then
+        logic.finish_race()
+        return
     end
 
-    -- Draw a card
-    local drawn_card
-    repeat
-        -- If the deck is empty, rebuild and reshuffle
-        if #current_state.deck == 0 then
-            driftwood.log.info("Deck is empty. Rebuilding and reshuffling the deck.")
-            racers.build_deck(current_state)
-
-            current_state.redraws = current_state.redraws + 1
-            if current_state.redraws >= constants.MAX_REDRAW_ATTEMPTS then
-                driftwood.log.info("Max redraw limit reached. Stopping the race.")
-                finishing = true
-            end
-        end
-
-        drawn_card = table.remove(current_state.deck, 1)
-
-        -- If the card belongs to a finished snail, discard and redraw
-        if drawn_card and current_state.positions[drawn_card.snail_id] >= constants.MAX_UNITS then
-            drawn_card = nil
-        end
-    until drawn_card
-
+    -- Draw a valid card
+    local drawn_card = draw_valid_card(current_state)
     if not drawn_card then
         driftwood.log.info("No valid cards left. Ending race.")
         logic.finish_race()
@@ -79,9 +86,9 @@ logic.race_tick = function()
         driftwood.log.error("Card not found: " .. drawn_card.card)
         return
     end
-    driftwood.log.info("Snail " .. drawn_card.snail_id .. " drew card: " .. card.name)
 
     -- Apply all effects from the card
+    driftwood.log.info("Snail " .. drawn_card.snail_id .. " drew card: " .. card.name)
     handlers.process_card(card.effects, drawn_card.snail_id, current_state)
 
     -- Update positions and check if any snail finishes
@@ -102,7 +109,7 @@ logic.race_tick = function()
     state.set(current_state)
 
     -- Check if all snails have finished
-    if all_snails_finished(current_state.positions) or finishing then
+    if all_snails_finished(current_state.positions) then
         logic.finish_race()
         return
     end
